Change basemap data to comply with CRAN
To comply with CRAN policy, the high resolution basemap originally distributed with mapmixture v1.1.1 has been replaced with the medium resolution basemap distributed with the `rnaturalearthdata` package, which `mapmixture` now imports as default for the `mapmixture()` function. See the `mapmixture` GitHub example page for the single line of code needed to import the high resolution basemap from the `rnaturalearthhires` package.